Split up Melua
Spread out functions over multiple files an cleaned up a little.
