--- src/Melua/LuaLib.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace MeluaLib
-{
-	public enum LuaLib
-	{
-		Base,
-		BaseSafe,
-		Package,
-		Table,
-		IO,
-		OS,
-		String,
-		Math,
-		Debug,
-	}
-}
